create methods list create update getbyid parts service

+++ src/cars/cars.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  ConflictException,
+} from '@nestjs/common';
 import { CreateCarDto } from './dto/create-car.dto';
 import { UpdateCarDto } from './dto/update-car.dto';
 import { Car } from './entities/car.entity';
@@ -13,7 +17,15 @@ export class CarsService {
     private carRepository: Repository<Car>,
   ) {}
 
-  create(createCarDto: CreateCarDto, id: string) {
+  async create(createCarDto: CreateCarDto, id: string, license_plate: string) {
+    const checkCar = await this.carRepository.findOne({
+      where: { license_plate: license_plate },
+    });
+
+    if (!checkCar) {
+      throw new ConflictException('Car with this license plate already exist');
+    }
+
     const carData: Car = {
       ...createCarDto,
       id: undefined,
@@ -72,8 +84,13 @@ export class CarsService {
     });
   }
 
-  findOne(id: string) {
-    return this.carRepository.findOne({ where: { id: id } });
+  async findOne(id: string) {
+    const checkCar = await this.carRepository.findOne({ where: { id: id } });
+    if (!checkCar) {
+      throw new NotFoundException('Car not found');
+    }
+
+    return checkCar;
   }
 
   async update(id: string, updateCarDto: UpdateCarDto) {
@@ -81,13 +98,21 @@ export class CarsService {
       where: { id },
     });
 
+    if (!carData) {
+      throw new NotFoundException('Can not find a car to update');
+    }
+
     return this.carRepository.save({
       ...carData,
       ...updateCarDto,
     });
   }
 
-  remove(id: string) {
+  async remove(id: string) {
+    const checkCar = await this.carRepository.findOne({ where: { id: id } });
+    if (!checkCar) {
+      throw new NotFoundException('Can not find a car to delete');
+    }
     return this.carRepository.delete({ id: id });
   }
 }
